+++ modules/cmsadmin/apis/CatController.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace cmsadmin\apis;
-
-class CatController extends \admin\ngrest\base\Api
-{
-    public $modelClass = 'cmsadmin\models\Cat';
-}