+++ modules/cmsadmin/apis/NavContainerController.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace cmsadmin\apis;
+
+class NavContainerController extends \admin\ngrest\base\Api
+{
+    public $modelClass = 'cmsadmin\models\NavContainer';
+}