+++ modules/cmsadmin/controllers/CatController.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace cmsadmin\controllers;
-
-class CatController extends \admin\ngrest\base\Controller
-{
-    public $modelClass = 'cmsadmin\models\Cat';
-}