+++ modules/cmsadmin/controllers/NavcontainerController.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace cmsadmin\controllers;
+
+class NavcontainerController extends \admin\ngrest\base\Controller
+{
+    public $modelClass = 'cmsadmin\models\NavContainer';
+}