+++ modules/cmsadmin/migrations/m141203_143059_cms_nav.php
@@ -8,7 +8,7 @@ class m141203_143059_cms_nav extends Migration
     {
         $this->createTable('cms_nav', [
             'id' => 'pk',
-            'cat_id' => 'int(11) NOT NULL DEFAULT 0',
+            'cat_id' => 'int(11) NOT NULL DEFAULT 0', // renamed to nav_container_id
             'parent_nav_id' => 'int(11) NOT NULL DEFAULT 0',
             'sort_index' => 'int(11) NOT NULL DEFAULT 0',
             'is_deleted' => 'tinyint(1) DEFAULT 0',